+++ pkg/network/message/message.go
@@ -13,6 +13,7 @@ import (
 
 const (
 	cWorkSizeKey = 1
+	cPuzzleIterN = 2
 
 	// third digits of PI
 	cAuthSalt = "8214808651_3282306647_0938446095_5058223172_5359408128"
@@ -30,7 +31,7 @@ type sMessage struct {
 
 func NewMessage(pSett ISettings, pPld payload.IPayload) IMessage {
 	hash := getHash(pSett.GetNetworkKey(), pPld.ToBytes())
-	proof := puzzle.NewPoWPuzzle(pSett.GetWorkSizeBits()).ProofBytes(hash)
+	proof := puzzle.NewPoWPuzzle(pSett.GetWorkSizeBits(), cPuzzleIterN).ProofBytes(hash)
 
 	return &sMessage{
 		fProof:   proof,
@@ -63,7 +64,7 @@ func LoadMessage(pSett ISettings, pData interface{}) (IMessage, error) {
 	copy(proofArray[:], proofBytes[:])
 
 	proof := encoding.BytesToUint64(proofArray)
-	puzzle := puzzle.NewPoWPuzzle(pSett.GetWorkSizeBits())
+	puzzle := puzzle.NewPoWPuzzle(pSett.GetWorkSizeBits(), cPuzzleIterN)
 	if !puzzle.VerifyBytes(gotHash, proof) {
 		return nil, errors.New("got invalid proof of work")
 	}